Ajustes semanticos realizados en la carpeta Domain y DTOs | correciones en los repositorios y ajustes en las clases services (sin código aún)

+++ backend/src/main/java/com/jhanp/gestioneventos/domain/view/ActiveEventsView.java
@@ -14,9 +14,8 @@ import java.time.LocalTime;
 @Getter
 @NoArgsConstructor
 public class ActiveEventsView {
-
     @Id
-    private Long id;
+    private Integer id;
 
     @Column(name = "nombre")
     private String name;
@@ -55,3 +54,4 @@ public class ActiveEventsView {
     private String organization;
 
 }
+
