+++ backend/src/main/java/com/jhanp/gestioneventos/dto/request/UserEventRoleRequestDTO.java
@@ -8,7 +8,8 @@ import lombok.*;
 @AllArgsConstructor
 @Builder
 public class UserEventRoleRequestDTO {
-    private Long idUser;
-    private Long idEvent;
-    private Long idEventRole;
+    private Integer idUser;
+    private Integer idEvent;
+    private Integer idEventRole;
 }
+
