+++ backend/src/main/java/com/jhanp/gestioneventos/repository/entity/EventAccessRepository.java
@@ -10,14 +10,15 @@ import org.springframework.stereotype.Repository;
 import java.math.BigDecimal;
 
 @Repository
-public interface EventAccessRepository extends JpaRepository <EventAccess, Long> {
+public interface EventAccessRepository extends JpaRepository <EventAccess, Integer> {
 
     // Procedimientos
     @Procedure(procedureName = "sp_crear_acceso_evento")
     void createEventAccess(@Param("p_evento_id") Integer idEvent, @Param("p_tipo_acceso_id") Integer idAccessType,
-            @Param("p_precio") BigDecimal price,@Param("p_cupos") Integer spots);
+            @Param("p_precio") BigDecimal price, @Param("p_cupos") Integer spots);
 
     // Funciones
     @Query(value = "SELECT fn_cupos_disponibles(:eventId)", nativeQuery = true)
     Integer getAvailableSpots(@Param("eventId") Integer idEvent);
 }
+
