+++ backend/src/main/java/com/jhanp/gestioneventos/repository/entity/EventRepository.java
@@ -8,9 +8,10 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EventRepository extends JpaRepository <Event, Long> {
+public interface EventRepository extends JpaRepository <Event, Integer> {
 
     // Funciones
     @Query(value = "SELECT fn_porcentaje_ocupacion(:eventId)", nativeQuery = true)
     Double getOccupationPercentage(@Param("eventId") Integer idEvent);
 }
+
