+++ backend/src/main/java/com/jhanp/gestioneventos/repository/entity/RoleSystemRepository.java
@@ -5,5 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RoleSystemRepository extends JpaRepository <SystemRole, Long> {
+public interface RoleSystemRepository extends JpaRepository <SystemRole, Integer> {
 }
+
